New way of handling scenario and scenario components
Now Scenario are used as a means to configure the environment
to run the set of tests. Scenario can be composed of several
scenario components as before.

This allows to run tests in different configuration, e.g. for
galera: simple deployment, deployment with pcmk_host_map,
(future) deployment with ipv6...